Moving to renderPass
